+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - [PopupElement](src/popup/popupElement.ts). Fix 2nd $show() has predefined maxSize & position can be different
 - [PopupElement](src/popup/popupElement.ts). Now popup has max size 100vw & 100vh (by default). Earlier popup could be > 100vw
 - [PopupElement](src/popup/popupElement.ts). No-updates on screensize changes. Now popup position & size changes with device-rotation
+- [PopupElement](src/popup/popupElement.ts). No popup if mousedown>move>up>click on target
 
 - **New**: